// frontend/pharmacy-app/src/enums/tableColumnKeys.ts
export enum CmmOrderTableColumnKeys {
  PATIENT_DETAILS = "patientDetails",
  DATE_OF_SERVICE = "dateOfService",
  DRUG = "drug",
  MEDICATION = "medication",
  PROVIDER_DETAILS = "providerDetails",
  PRIMARY_DIAGNOSES_DESCRIPTION = "primaryDiagnosesDescription",
  DATE_OF_BIRTH = "dateOfBirth",
  INSURANCE_STATE = "insuranceState",
  ACTION = "button",
  STATUS = "status",
  OUTCOME = "outcome",
  CMM_KEY = "cmmKey",
  FORM_NAME = "formName",
  ID = "id",
  DELETE_ICON = "cmmOrderDeleteIcon",
  PATIENT_ID = "patientId",
  ROW_DATA = "rowData",
  NO_DATA_FIELDS = "noDataFields",
  USERNAME = "username",
  BUTTON_WITH_THREE_DOTS = "buttonWithThreeDots",
  CMM_INPUT_VIEW_ICON = "cmmInputViewIcon",
  TYPE = "type",
  ADDITIONAL_NOTES = "additionalNotes",
  KEY_POC = "keyPoc",
  ISSUE = "issue",
  RECORD_CLOSED_BY = "recordClosedBy",
}